local-llm-vlm-experiments/requirements.txt
transformers>=4.35.0
torch>=2.0.0
ollama>=0.1.0
python-dotenv>=1.0.0
bitsandbytes>=0.41.0
accelerate>=0.24.0
pillow>=10.0.0
